refactor code with different folder for prompt and moving interface to adapter
