+++ vscode/src/autoedits/adapters/utils.ts
@@ -45,3 +45,26 @@ export function getSourcegraphCompatibleChatPrompt(param: {
     prompt.push({ speaker: 'human', text: param.userMessage })
     return prompt
 }
+
+export async function getModelResponse(
+    url: string,
+    body: string,
+    apiKey: string,
+    customHeaders: Record<string, string> = {}
+): Promise<any> {
+    const response = await fetch(url, {
+        method: 'POST',
+        headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${apiKey}`,
+            ...customHeaders,
+        },
+        body: body,
+    })
+    if (response.status !== 200) {
+        const errorText = await response.text()
+        throw new Error(`HTTP error! status: ${response.status}, message: ${errorText}`)
+    }
+    const data = await response.json()
+    return data
+}
