+++ vscode/src/autoedits/autoedits-provider.ts
@@ -19,20 +19,16 @@ import { DefaultContextStrategyFactory } from '../completions/context/context-st
 import { RetrieverIdentifier } from '../completions/context/utils'
 import { getCurrentDocContext } from '../completions/get-current-doc-context'
 import { getConfiguration } from '../configuration'
-
 import { CodyGatewayAdapter } from './adapters/cody-gateway'
 import { FireworksAdapter } from './adapters/fireworks'
 import { OpenAIAdapter } from './adapters/openai'
 import { SourcegraphChatAdapter } from './adapters/sourcegraph-chat'
 import { SourcegraphCompletionsAdapter } from './adapters/sourcegraph-completions'
 import { autoeditsLogger } from './logger'
-import type { AutoeditsModelAdapter, AutoeditsPrompt, PromptResponseData } from './prompt-provider'
-import {
-    type CodeToReplaceData,
-    SYSTEM_PROMPT,
-    getBaseUserPrompt,
-    getCompletionsPromptWithSystemPrompt,
-} from './prompt-utils'
+import type { AutoeditsUserPromptStrategy } from './prompt/base'
+import { SYSTEM_PROMPT } from './prompt/constants'
+import { DefaultUserPromptStrategy } from './prompt/default-prompt-strategy'
+import { type CodeToReplaceData, getCompletionsPromptWithSystemPrompt } from './prompt/prompt-utils'
 import { DefaultDecorator } from './renderer/decorators/default-decorator'
 import { InlineDiffDecorator } from './renderer/decorators/inline-diff-decorator'
 import { getDecorationInfo } from './renderer/diff-utils'
@@ -83,6 +79,7 @@ export class AutoeditsProvider implements vscode.InlineCompletionItemProvider, v
     private readonly onSelectionChangeDebounced: DebouncedFunc<typeof this.autoeditOnSelectionChange>
     /** Keeps track of the last time the text was changed in the editor. */
     private lastTextChangeTimeStamp: number | undefined
+    private readonly promptProvider: AutoeditsUserPromptStrategy = new DefaultUserPromptStrategy()
 
     private isMockResponseFromCurrentDocumentTemplateEnabled = vscode.workspace
         .getConfiguration()
@@ -325,14 +322,17 @@ export class AutoeditsProvider implements vscode.InlineCompletionItemProvider, v
         position: vscode.Position,
         context: AutocompleteContextSnippet[],
         tokenBudget: AutoEditsTokenLimit
-    ): PromptResponseData {
-        const { codeToReplace, prompt: userPrompt } = getBaseUserPrompt(
+    ): {
+        codeToReplace: CodeToReplaceData
+        promptResponse: AutoeditsPrompt
+    } {
+        const { codeToReplace, prompt: userPrompt } = this.promptProvider.getUserPrompt({
             docContext,
             document,
             position,
             context,
-            tokenBudget
-        )
+            tokenBudget,
+        })
         const prompt: AutoeditsPrompt = this.config.isChatModel
             ? { systemMessage: SYSTEM_PROMPT, userMessage: userPrompt }
             : { userMessage: getCompletionsPromptWithSystemPrompt(SYSTEM_PROMPT, userPrompt) }
